Removes version dropdown from home page
According to the current architecture of the repo, it is imposibble
to configure a default dropdown item for the home page. And, as long
as the Home page shouldn't be versioned, the suggestion is to remove
the version dropdown from it completely.

The change introduces the `hideVersionDropdown` page param to conditionally
hide version selector:

- Introduced `hideVersionDropdown` front matter parameter for pages
- Wrap version dropdown in `navbar.html` with the `if not .Params.hideVersionDropdown`